+++ fem/fespace.hpp
@@ -925,229 +925,6 @@ public:
    void Save(std::ostream &out) const;
 };
 
-
-/** @brief Base class for transfer algorithms that construct transfer Operator%s
-    between two finite element (FE) spaces. */
-/** Generally, the two FE spaces (domain and range) can be defined on different
-    meshes. */
-class GridTransfer
-{
-protected:
-   FiniteElementSpace &dom_fes; ///< Domain FE space
-   FiniteElementSpace &ran_fes; ///< Range FE space
-
-   /** @brief Desired Operator::Type for the construction of all operators
-       defined by the underlying transfer algorithm. It can be ignored by
-       derived classes. */
-   Operator::Type oper_type;
-
-   OperatorHandle fw_t_oper; ///< Forward true-dof operator
-   OperatorHandle bw_t_oper; ///< Backward true-dof operator
-
-#ifdef MFEM_USE_MPI
-   bool parallel;
-#endif
-   bool Parallel() const
-   {
-#ifndef MFEM_USE_MPI
-      return false;
-#else
-      return parallel;
-#endif
-   }
-
-   const Operator &MakeTrueOperator(FiniteElementSpace &fes_in,
-                                    FiniteElementSpace &fes_out,
-                                    const Operator &oper,
-                                    OperatorHandle &t_oper);
-
-public:
-   /** Construct a transfer algorithm between the domain, @a dom_fes_, and
-       range, @a ran_fes_, FE spaces. */
-   GridTransfer(FiniteElementSpace &dom_fes_, FiniteElementSpace &ran_fes_);
-
-   /// Virtual destructor
-   virtual ~GridTransfer() { }
-
-   /** @brief Set the desired Operator::Type for the construction of all
-       operators defined by the underlying transfer algorithm. */
-   /** The default value is Operator::ANY_TYPE which typically corresponds to
-       a matrix-free operator representation. Note that derived classes are not
-       required to support this setting and can ignore it. */
-   void SetOperatorType(Operator::Type type) { oper_type = type; }
-
-   /** @brief Return an Operator that transfers GridFunction%s from the domain
-       FE space to GridFunction%s in the range FE space. */
-   virtual const Operator &ForwardOperator() = 0;
-
-   /** @brief Return an Operator that transfers GridFunction%s from the range
-       FE space back to GridFunction%s in the domain FE space. */
-   virtual const Operator &BackwardOperator() = 0;
-
-   /** @brief Return an Operator that transfers true-dof Vector%s from the
-       domain FE space to true-dof Vector%s in the range FE space. */
-   /** This method is implemented in the base class, based on ForwardOperator(),
-       however, derived classes can overload the construction, if necessary. */
-   virtual const Operator &TrueForwardOperator()
-   {
-      return MakeTrueOperator(dom_fes, ran_fes, ForwardOperator(), fw_t_oper);
-   }
-
-   /** @brief Return an Operator that transfers true-dof Vector%s from the
-       range FE space back to true-dof Vector%s in the domain FE space. */
-   /** This method is implemented in the base class, based on
-       BackwardOperator(), however, derived classes can overload the
-       construction, if necessary. */
-   virtual const Operator &TrueBackwardOperator()
-   {
-      return MakeTrueOperator(ran_fes, dom_fes, BackwardOperator(), bw_t_oper);
-   }
-};
-
-
-/** @brief Transfer data between a coarse mesh and an embedded refined mesh
-    using interpolation. */
-/** The forward, coarse-to-fine, transfer uses nodal interpolation. The
-    backward, fine-to-coarse, transfer is defined locally (on a coarse element)
-    as B = (F^t M_f F)^{-1} F^t M_f, where F is the forward transfer matrix, and
-    M_f is a mass matrix on the union of all fine elements comprising the coarse
-    element. Note that the backward transfer operator, B, is a left inverse of
-    the forward transfer operator, F, i.e. B F = I. Both F and B are defined in
-    reference space and do not depend on the actual physical shape of the mesh
-    elements.
-
-    It is assumed that both the coarse and the fine FiniteElementSpace%s use
-    compatible types of elements, e.g. finite elements with the same map-type
-    (VALUE, INTEGRAL, H_DIV, H_CURL - see class FiniteElement). Generally, the
-    FE spaces can have different orders, however, in order for the backward
-    operator to be well-defined, the (local) number of the fine dofs should not
-    be smaller than the number of coarse dofs. */
-class InterpolationGridTransfer : public GridTransfer
-{
-protected:
-   BilinearFormIntegrator *mass_integ; ///< Ownership depends on #own_mass_integ
-   bool own_mass_integ; ///< Ownership flag for #mass_integ
-
-   OperatorHandle F; ///< Forward, coarse-to-fine, operator
-   OperatorHandle B; ///< Backward, fine-to-coarse, operator
-
-public:
-   InterpolationGridTransfer(FiniteElementSpace &coarse_fes,
-                             FiniteElementSpace &fine_fes)
-      : GridTransfer(coarse_fes, fine_fes),
-        mass_integ(NULL), own_mass_integ(false)
-   { }
-
-   virtual ~InterpolationGridTransfer();
-
-   /** @brief Assign a mass integrator to be used in the construction of the
-       backward, fine-to-coarse, transfer operator. */
-   void SetMassIntegrator(BilinearFormIntegrator *mass_integ_,
-                          bool own_mass_integ_ = true);
-
-   virtual const Operator &ForwardOperator();
-
-   virtual const Operator &BackwardOperator();
-};
-
-
-/** @brief Transfer data between a coarse mesh and an embedded refined mesh
-    using L2 projection. */
-/** The forward, coarse-to-fine, transfer uses L2 projection. The backward,
-    fine-to-coarse, transfer is defined locally (on a coarse element) as
-    B = (F^t M_f F)^{-1} F^t M_f, where F is the forward transfer matrix, and
-    M_f is the mass matrix on the union of all fine elements comprising the
-    coarse element. Note that the backward transfer operator, B, is a left
-    inverse of the forward transfer operator, F, i.e. B F = I. Both F and B are
-    defined in physical space and, generally, vary between different mesh
-    elements.
-
-    This class currently only fully supports L2 finite element spaces and fine
-    meshes that are a uniform refinement of the coarse mesh. Generally, the
-    coarse and fine FE spaces can have different orders, however, in order for
-    the backward operator to be well-defined, the number of the fine dofs (in a
-    coarse element) should not be smaller than the number of coarse dofs.
-
-    If used on H1 finite element spaces, the transfer will be performed locally,
-    and the value of shared (interface) degrees of freedom will be determined by
-    the value of the last transfer to be performed (according to the element
-    numbering in the finite element space). As a consequence, the mass
-    conservation properties for this operator from the L2 case do not carry over
-    to H1 spaces. */
-class L2ProjectionGridTransfer : public GridTransfer
-{
-protected:
-   /** Class representing projection operator between a high-order L2 finite
-       element space on a coarse mesh, and a low-order L2 finite element space
-       on a refined mesh (LOR). We assume that the low-order space, fes_lor,
-       lives on a mesh obtained by refining the mesh of the high-order space,
-       fes_ho. */
-   class L2Projection : public Operator
-   {
-      const FiniteElementSpace &fes_ho;
-      const FiniteElementSpace &fes_lor;
-
-      // The restriction and prolongation operators are represented as dense
-      // elementwise matrices (of potentially different sizes, because of mixed
-      // meshes or p-refinement). The matrix entries are stored in the R and P
-      // arrays. The entries of the i'th high-order element are stored at the
-      // index given by offsets[i].
-      mutable Array<double> R, P;
-      Array<int> offsets;
-
-      Table ho2lor;
-   public:
-      L2Projection(const FiniteElementSpace &fes_ho_,
-                   const FiniteElementSpace &fes_lor_);
-      /// Perform the L2 projection onto the LOR space
-      virtual void Mult(const Vector &x, Vector &y) const;
-      /// Perform the transpose of L2 projection onto the LOR space, useful for
-      /// transferring dual fields.
-      virtual void MultTranspose(const Vector &x, Vector &y) const;
-      /// Perform the mass conservative left-inverse prolongation operation.
-      /// This functionality is also provided as an Operator by L2Prolongation.
-      void Prolongate(const Vector &x, Vector &y) const;
-      /// Perform the transpose of the mass conservative left-inverse
-      /// prolongation operation, useful for transferring dual fields.
-      void ProlongateTranspose(const Vector &x, Vector &y) const;
-      virtual ~L2Projection() { }
-   };
-
-   /** Mass-conservative prolongation operator going in the opposite direction
-       as L2Projection. This operator is a left inverse to the L2Projection. */
-   class L2Prolongation : public Operator
-   {
-      const L2Projection &l2proj;
-
-   public:
-      L2Prolongation(const L2Projection &l2proj_)
-         : Operator(l2proj_.Width(), l2proj_.Height()), l2proj(l2proj_) { }
-      void Mult(const Vector &x, Vector &y) const
-      {
-         l2proj.Prolongate(x, y);
-      }
-      void MultTranspose(const Vector &x, Vector &y) const
-      {
-          l2proj.ProlongateTranspose(x, y);
-      }
-      virtual ~L2Prolongation() { }
-   };
-
-   L2Projection   *F; ///< Forward, coarse-to-fine, operator
-   L2Prolongation *B; ///< Backward, fine-to-coarse, operator
-
-public:
-   L2ProjectionGridTransfer(FiniteElementSpace &coarse_fes,
-                            FiniteElementSpace &fine_fes)
-      : GridTransfer(coarse_fes, fine_fes),
-        F(NULL), B(NULL)
-   { }
-
-   virtual const Operator &ForwardOperator();
-
-   virtual const Operator &BackwardOperator();
-};
-
 inline bool UsesTensorBasis(const FiniteElementSpace& fes)
 {
    return dynamic_cast<const mfem::TensorBasisElement *>(fes.GetFE(0))!=nullptr;
